factory v2
split the code into different classes and new dirs
