+++ BattleshipMpClient/BattleshipMpClient.csproj
@@ -47,6 +47,15 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Client.cs" />
+    <Compile Include="Factory\Ship\Battleship.cs" />
+    <Compile Include="Factory\Ship\Cruiser.cs" />
+    <Compile Include="Factory\Ship\Destroyer.cs" />
+    <Compile Include="Factory\Ship\IShip.cs" />
+    <Compile Include="Factory\Ship\IShipFactory.cs" />
+    <Compile Include="Factory\Ship\ShipButtons.cs" />
+    <Compile Include="Factory\Ship\ShipFactory.cs" />
+    <Compile Include="Factory\Ship\ShipType.cs" />
+    <Compile Include="Factory\Ship\Submarine.cs" />
     <Compile Include="Form1_ClientScreen.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -73,7 +82,6 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Ship.cs" />
     <EmbeddedResource Include="Form1_ClientScreen.resx">
       <DependentUpon>Form1_ClientScreen.cs</DependentUpon>
     </EmbeddedResource>
