+++ BattleshipMpClient/Factory/Ship/Battleship.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BattleshipMpClient.Factory.Ship
+{
+    public class Battleship : IShip
+    {
+        public string shipName => "Battleship";
+        public int remShips { get; set; } = 1;
+        public List<ShipButtons> shipPerButton { get; set; } = new List<ShipButtons>();
+    }
+}
