+++ BattleshipMpClient/Factory/Ship/Destroyer.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BattleshipMpClient.Factory.Ship
+{
+    public class Destroyer : IShip
+    {
+        public string shipName => "Destroyer";
+        public int remShips { get; set; } = 3;
+        public List<ShipButtons> shipPerButton { get; set; } = new List<ShipButtons>();
+    }
+}
