+++ BattleshipMpClient/Factory/Ship/ShipButtons.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BattleshipMpClient.Factory.Ship
+{
+    public class ShipButtons
+    {
+        public List<string> buttonNames { get; set; } = new List<string>();
+    }
+}
