+++ BattleshipMpServer/BattleshipMpServer.csproj
@@ -46,6 +46,15 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Factory\Ship\Battleship.cs" />
+    <Compile Include="Factory\Ship\Cruiser.cs" />
+    <Compile Include="Factory\Ship\Destroyer.cs" />
+    <Compile Include="Factory\Ship\IShip.cs" />
+    <Compile Include="Factory\Ship\IShipFactory.cs" />
+    <Compile Include="Factory\Ship\ShipButtons.cs" />
+    <Compile Include="Factory\Ship\ShipFactory.cs" />
+    <Compile Include="Factory\Ship\ShipType.cs" />
+    <Compile Include="Factory\Ship\Submarine.cs" />
     <Compile Include="Form1_ServerScreen.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -73,7 +82,6 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Server.cs" />
-    <Compile Include="Ship.cs" />
     <EmbeddedResource Include="Form1_ServerScreen.resx">
       <DependentUpon>Form1_ServerScreen.cs</DependentUpon>
     </EmbeddedResource>
