+++ BattleshipMpServer/Factory/Ship/Cruiser.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BattleshipMpServer.Factory.Ship
+{
+    public class Cruiser : IShip
+    {
+        public string shipName => "Cruiser";
+        public int remShips { get; set; } = 2;
+        public List<ShipButtons> shipPerButton { get; set; } = new List<ShipButtons>();
+    }
+}
