+++ BattleshipMpServer/Factory/Ship/IShip.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BattleshipMpServer.Factory.Ship
+{
+    public interface IShip
+    {
+        string shipName { get; }
+        int remShips { get; set; }
+        List<ShipButtons> shipPerButton { get; set; }
+    }
+}
