+++ BattleshipMpServer/Factory/Ship/ShipType.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BattleshipMpServer.Factory.Ship
+{
+    public enum ShipType
+    {
+        Battleship,
+        Cruiser,
+        Destroyer,
+        Submarine
+    }
+}
