+++ BattleshipMpServer/Form2_PreparatoryScreen.cs
@@ -10,6 +10,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using BattleshipMpServer.Factory.Ship;
 
 namespace BattleshipMp
 {
@@ -91,11 +92,12 @@ namespace BattleshipMp
 
         #endregion
 
-        public static List<string> shipTypes = new List<string>()
+        private static readonly ShipType[] shipTypes =
         {
-            "Battleship", "Cruiser", "Destroyer", "Submarine"
+            ShipType.Battleship, ShipType.Cruiser, ShipType.Destroyer, ShipType.Submarine
         };
 
+        IShipFactory shipFactory = new ShipFactory();
         public static List<IShip> shipList = new List<IShip>();
 
         List<string> AllSelectedButtonList = new List<string>();
@@ -120,7 +122,7 @@ namespace BattleshipMp
             }
             foreach (var shipType in shipTypes)
             {
-                IShip _ship = ShipFactory.CreateShip(shipType);
+                IShip _ship = shipFactory.CreateShip(shipType);
 
                 shipList.Add(_ship);
             }
