+++ BattleshipMpServer/Form4_GameScreen.cs
@@ -10,6 +10,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using BattleshipMpServer.Factory.Ship;
 using static System.Windows.Forms.VisualStyles.VisualStyleElement.Button;
 
 namespace BattleshipMp
